apanhando dos horarios na main
