+++ folhaPagamento/_Classes/Registro.cs
@@ -1,5 +1,7 @@
-﻿using System;
+﻿using folhaPagamento._DAO;
+using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -18,5 +20,10 @@ namespace folhaPagamento._Classes
     {
         public const string carregaRegistro = "SELECT * FROM ponto;";
         public const string adicionaRegistro = "INSERT INTO ponto (cpf_ponto, data, hora) VALUES (@cpf_ponto, @data, @hora);";
+        public static DataTable FiltrarRegistros(string filtroRegistro)
+        {
+            string consultaRegistro = "SELECT * FROM ponto WHERE cpf_ponto LIKE '%" + filtroRegistro + "%'";
+            return PontoDAO.ExecutarConsulta(consultaRegistro);
+        }
     }
 }
