+++ folhaPagamento/_Forms/MainWF.cs
@@ -13,100 +13,17 @@ namespace folhaPagamento
     {
         private User usuarios;
         public Funcionario Usuarios { get; set; }
+        Principal principal = new Principal();
 
-        private Label labelSegunda;
-        private Label labelTerca;
-        private Label labelQuarta;
-        private Label labelQuinta;
-        private Label labelSexta;
-        private Label labelSabado;
-        private Label labelDomingo;
+
 
         public MainWF(Funcionario usuarios)
         {
             InitializeComponent();
             Usuarios = usuarios;
 
-            var dataAtual = DateTime.Today;
-            var cultura = CultureInfo.CurrentCulture;
-            const int corPainelDataAtual = 218;
-
-            int diasPassados = (int)dataAtual.DayOfWeek;
-            for (int i = 0; i < 7; i++)
-            {
-                DateTime data = dataAtual.AddDays(i - diasPassados);
-                string nomeDiaSemana = cultura.DateTimeFormat.GetDayName(data.DayOfWeek);
-                string diaMes = data.ToString("dd/MM");
-                string resultado = $"{nomeDiaSemana} ({diaMes})";
-
-                // Exibir o resultado em cada label correspondente
-                SetTextForLabelDayOfWeek(data.DayOfWeek, resultado);
-
-                // Definir o estilo de borda para o painel da data de hoje
-                if (data.Date == dataAtual)
-                {
-                    Panel panel = GetPanelForDayOfWeek(data.DayOfWeek);
-                    panel.BackColor = Color.FromArgb(corPainelDataAtual, corPainelDataAtual, corPainelDataAtual);
-                }
-            }
-
-
+            principal.carregaDiasSemana();
         }
-
-        private Panel GetPanelForDayOfWeek(DayOfWeek dayOfWeek)
-        {
-            switch (dayOfWeek)
-            {
-                case DayOfWeek.Monday:
-                    return panelSegunda;
-                case DayOfWeek.Tuesday:
-                    return panelTerca;
-                case DayOfWeek.Wednesday:
-                    return panelQuarta;
-                case DayOfWeek.Thursday:
-                    return panelQuinta;
-                case DayOfWeek.Friday:
-                    return panelSexta;
-                case DayOfWeek.Saturday:
-                    return panelSabado;
-                case DayOfWeek.Sunday:
-                    return panelDomingo;
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(dayOfWeek));
-            }
-        }
-
-        private void SetTextForLabelDayOfWeek(DayOfWeek dayOfWeek, string text)
-        {
-            switch (dayOfWeek)
-            {
-                case DayOfWeek.Monday:
-                    lblSegunda.Text = text;
-                    break;
-                case DayOfWeek.Tuesday:
-                    lblTerca.Text = text;
-                    break;
-                case DayOfWeek.Wednesday:
-                    lblQuarta.Text = text;
-                    break;
-                case DayOfWeek.Thursday:
-                    lblQuinta.Text = text;
-                    break;
-                case DayOfWeek.Friday:
-                    lblSexta.Text = text;
-                    break;
-                case DayOfWeek.Saturday:
-                    lblSabado.Text = text;
-                    break;
-                case DayOfWeek.Sunday:
-                    lblDomingo.Text = text;
-                    break;
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(dayOfWeek));
-            }
-        }
-
-
 
         private void btnConfig_Click(object sender, EventArgs e)
         {
@@ -121,25 +38,7 @@ namespace folhaPagamento
                 btnEmpresa.Visible = false;
                 btnMinhasConfig.Visible = false;
             }
-
-            lblSaudacao.Text = Saudacao(Usuarios.nome);
-        }
-
-        string Saudacao(string nome)
-        {
-            DateTimeOffset now = DateTimeOffset.Now;
-            if (now.Hour >= 5 && now.Hour < 12)
-            {
-                return $"Bom dia, {nome}.";
-            }
-            else if (now.Hour >= 12 && now.Hour < 18)
-            {
-                return $"Boa tarde, {nome}.";
-            }
-            else
-            {
-                return $"Boa noite, {nome}.";
-            }
+            lblSaudacao.Text = principal.Saudacao(Usuarios.nome);
         }
 
         private void btnEmpresa_Click(object sender, EventArgs e)
@@ -193,22 +92,7 @@ namespace folhaPagamento
 
         private void btnFazerMarcacao_Click(object sender, EventArgs e)
         {
-            Registro novoRegistro = new Registro()
-            {
-                cpf_ponto = Usuarios.cpf,
-                data = DateTime.Now.Date,
-                hora = TimeSpan.FromSeconds(Math.Floor(DateTime.Now.TimeOfDay.TotalSeconds))
-            };
-
-            string mensagemConfirmacao = $"Efetuar marcação para:\nNome: {Usuarios.nome}\nCPF: {novoRegistro.cpf_ponto}\nData: {novoRegistro.data.ToString("dd/MM/yyyy")}\nHora: {novoRegistro.hora.ToString(@"hh\:mm\:ss")}";
-            DialogResult resultado = MessageBox.Show(mensagemConfirmacao, "Revise as informações", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-
-            if (resultado == DialogResult.Yes)
-            {
-                PontoDAO pontoDAO = new PontoDAO();
-                pontoDAO.RegistrarPonto(novoRegistro.cpf_ponto, novoRegistro.data, novoRegistro.hora);
-            }
-
+            principal.FazerMarcacao();
         }
 
         private void btnMinhasConfig_Click(object sender, EventArgs e)
